Moved some NNS governance test code to its own file.
